feat: SearchBox Component
Designed the searchbox component, allowing us to compatmentalize the search input functionality.

This has the benefit of allowing us apply a debounce on the search input which is used to delay running the search function at every query change
and avoid making too many API requests concurently.

Of course the lodash.debounce package was also installed to include this debounce functionality.
Some other quality of life improvements were included, such as including an error message when no books are returned by the search query.

A bug was fixed which was caused by no imageLinks property and stoped the rendering of the shelf books, I had to check for the imageLinks peoperty
before rendering the book.